+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.5.48 (Pending):
 
 ## 0.5.47:
-* When a client-side session is disocnnected, delete the server-side session.
+* When a client-side session is disconnected, delete the server-side session.
 * Upgrade to compile and test against GWT 2.7.0.
 * Replace TokenRestService with enhanced SessionRestService that provides status details
   about the session and other candidate sessions.